feat: pluggable strategy architecture (Phase 1)
- Add pkg/strategy/ public API package:
  - Strategy interface, Signal, MarketData, StrategyMetrics, StrategyStatus
  - BaseStrategy embed helper
  - Registry (replaces hardcoded StrategyFactory)
  - pkg/strategy/scalping/ reference implementation

- Add pkg/engine/ Run() API:
  - engine.Run(ctx, opts...) wires the full application
  - engine.WithStrategy() / WithConfigPath() functional options
  - TradingController, appServiceAdapter, marketDataAdapter moved from cmd/ to engine

- Update internal packages to import pkg/strategy (not internal/usecase/strategy):
  - internal/adapter/worker/{strategy_worker,signal_worker,strategy_monitor}
  - internal/adapter/http/contracts
  - internal/usecase/risk/manager

- Fix logger concrete type in trading sub-packages:
  - bitflyer_trader.go and monitor/order_monitor.go now accept logger.LoggerInterface

- Remove superseded files:
  - internal/usecase/strategy/ package del  - internal/usecase/strategy/ package del  - internal/usecase/strategy/ package del  - internal/  -  - internal/usecase/strategy/ packageategy()
  - Calls engine.Run() — no config/logger setup in main
