+++ internal/adapter/http/contracts.go
@@ -4,7 +4,7 @@ import (
 	"context"
 
 	"github.com/bmf-san/gogocoin/v1/internal/domain"
-	"github.com/bmf-san/gogocoin/v1/internal/usecase/strategy"
+	strategy "github.com/bmf-san/gogocoin/v1/pkg/strategy"
 )
 
 // DatabaseService defines the database operations needed by the API server.
