+++ internal/adapter/worker/signal_worker.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/bmf-san/gogocoin/v1/internal/domain"
 	"github.com/bmf-san/gogocoin/v1/internal/logger"
-	"github.com/bmf-san/gogocoin/v1/internal/usecase/strategy"
+	strategy "github.com/bmf-san/gogocoin/v1/pkg/strategy"
 )
 
 // SignalWorker processes trading signals and executes trades
