+++ internal/adapter/worker/strategy_monitor.go
@@ -5,7 +5,7 @@ import (
 	"time"
 
 	"github.com/bmf-san/gogocoin/v1/internal/logger"
-	"github.com/bmf-san/gogocoin/v1/internal/usecase/strategy"
+	strategy "github.com/bmf-san/gogocoin/v1/pkg/strategy"
 )
 
 // StrategyMonitorWorker monitors strategy health
