+++ internal/adapter/worker/strategy_worker.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/bmf-san/gogocoin/v1/internal/domain"
 	"github.com/bmf-san/gogocoin/v1/internal/logger"
-	"github.com/bmf-san/gogocoin/v1/internal/usecase/strategy"
+	strategy "github.com/bmf-san/gogocoin/v1/pkg/strategy"
 )
 
 const (
