+++ internal/usecase/risk/manager_test.go
@@ -6,7 +6,7 @@ import (
 	"time"
 
 	"github.com/bmf-san/gogocoin/v1/internal/domain"
-	"github.com/bmf-san/gogocoin/v1/internal/usecase/strategy"
+	strategy "github.com/bmf-san/gogocoin/v1/pkg/strategy"
 )
 
 // Mock implementations
