+++ pkg/strategy/scalping/params.go
@@ -1,0 +1,25 @@
+package scalping
+
+// Params holds all configuration for the scalping strategy.
+type Params struct {
+	EMAFastPeriod  int                       `yaml:"ema_fast_period"`
+	EMASlowPeriod  int                       `yaml:"ema_slow_period"`
+	TakeProfitPct  float64                   `yaml:"take_profit_pct"`
+	StopLossPct    float64                   `yaml:"stop_loss_pct"`
+	CooldownSec    int                       `yaml:"cooldown_sec"`
+	MaxDailyTrades int                       `yaml:"max_daily_trades"`
+	MinNotional    float64                   `yaml:"min_notional"`
+	FeeRate        float64                   `yaml:"fee_rate"`
+	RSIPeriod      int                       `yaml:"rsi_period"`
+	RSIOverbought  float64                   `yaml:"rsi_overbought"`
+	RSIOversold    float64                   `yaml:"rsi_oversold"`
+	SymbolParams   map[string]SymbolOverride `yaml:"symbol_params"`
+}
+
+// SymbolOverride holds per-symbol overrides; zero values fall back to global defaults.
+type SymbolOverride struct {
+	EMAFastPeriod int     `yaml:"ema_fast_period"`
+	EMASlowPeriod int     `yaml:"ema_slow_period"`
+	CooldownSec   int     `yaml:"cooldown_sec"`
+	MinNotional   float64 `yaml:"min_notional"`
+}
